requirements.txt
torch
torchvision
numpy
matplotlib
Pillow
tqdm
